WIP for new constraint handling
This tries to unify the behavior by treating new constraint the same as
PublicParameterlessConstructor annotation. But that doesn't work because
the annotation warnings fall into the reflection category, while the
new constraint warnings don't.

For example, the new constraint should warns when the substituted type
argument has RDC. But DAM.PublicParameterlessConstructor should not,
if trim analysis is disabled.